add progress and tag(tagstyle)

+++ mdvn-common/src/main/java/com/mdvns/mdvn/common/beans/RequirementInfo.java
@@ -42,7 +42,7 @@ public class RequirementInfo {
     /* requirement RAG status, ie. Red, Amber, Green*/
     private String ragStatus;
     /* requirement progress*/
-    private Double progress;
+    private Float progress;
     /* total amount of story point and CR story point*/
     private Integer totalStoryPoint;
     //备注
@@ -167,11 +167,11 @@ public class RequirementInfo {
         this.ragStatus = ragStatus;
     }
 
-    public Double getProgress() {
+    public Float getProgress() {
         return progress;
     }
 
-    public void setProgress(Double progress) {
+    public void setProgress(Float progress) {
         this.progress = progress;
     }
